river.genetic/requirements.txt
appnope==0.1.0
backports-abc==0.4
boto==2.43.0
cycler==0.10.0
decorator==4.0.10
entrypoints==0.2.2
ipykernel==4.5.0
ipython==5.1.0
ipython-genutils==0.1.0
ipywidgets==5.2.2
Jinja2==2.8
jsonschema==2.5.1
jupyter-client==4.4.0
jupyter-core==4.2.0
MarkupSafe==0.23
matplotlib==1.5.3
mistune==0.7.3
nbconvert==4.2.0
nbformat==4.1.0
notebook==4.2.3
numpy==1.11.2
parse==1.6.6
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.8
ptyprocess==0.5.1
py4j==0.10.4
Pygments==2.1.3
pyparsing==2.1.10
pysparkling==0.3.24
python-dateutil==2.5.3
pytz==2016.7
pyzmq==16.0.0
requests==2.11.1
simplegeneric==0.8.1
six==1.10.0
terminado==0.6
tornado==4.4.2
traitlets==4.3.1
wcwidth==0.1.7
widgetsnbextension==1.2.6
